requirements.txt
Django==4.2
djangorestframework==3.14.0
django-ckeditor==6.7.0
redis==4.6.0
googletrans==4.0.0-rc1
pytest==7.4.0
pytest-django==4.7.0
flake8==6.1.0